According to kernel sources, that is XNU 344.49 (10.2.8), XNU 517.12.7 (10.3.9), XNU 792.21.3 (10.4.10) and XNU 1228 (10.5.0), exception handler code[1] always contains the fault address nowadays. So make it the default fast path but keep provisions to check that at run-time first.
This yields a nearly 4x improvement in SIGSEGV recovery but MacOS X is still
suboptimal wrt. Linux, so VOSF is still not possible with frameskip == 0.

XXX: the ppc kernel had bugs that caused DAR (put into code[1]) to be incorrectly
decoded. This would need a broader test audience or more careful audit of the
sources changes.